refactor: content code blocks
